Rework InteractionBlock
The implementation from this commit is suboptimal in many ways, but it
is simple and allowed me to update the unit tests with few minor
changes.
The current implementation will most likely be replaced in the near
future, but no changes to the test cases should be required (in theory).

+++ src/hic/include/hictk/hic/hic_matrix_selector.hpp
@@ -43,7 +43,7 @@ class MatrixSelector {
   [[nodiscard]] const Chromosome &chrom1() const noexcept;
   [[nodiscard]] const Chromosome &chrom2() const noexcept;
 
-  [[nodiscard]] std::int64_t resolution() const noexcept;
+  [[nodiscard]] std::uint32_t resolution() const noexcept;
   [[nodiscard]] MatrixType matrix_type() const noexcept;
   [[nodiscard]] NormalizationMethod normalizationMethod() const noexcept;
   [[nodiscard]] MatrixUnit matrixUnit() const noexcept;
